write custom hook for isShaking
