+++ App.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Accelerometer } from "expo-sensors";
 import styled from "styled-components/native";
+import { useIsShaking } from "./utils/utils";
 
 const Container = styled.View`
   flex: 1;
@@ -16,29 +17,7 @@ const Title = styled.Text`
 
 const App = () => {
   const [quote, setQuote] = useState("");
-  const [isShaking, setIsShaking] = useState(false);
-  const [subscription, setSubscription] = useState(null);
-
-  Accelerometer.setUpdateInterval(400);
-
-  const _subscribe = () => {
-    setSubscription(
-      Accelerometer.addListener((accelerometerData) => {
-        const { x, y, z } = accelerometerData;
-        setIsShaking(Math.abs(x) + Math.abs(y) + Math.abs(z) > 1.78);
-      })
-    );
-  };
-
-  const _unsubscribe = () => {
-    subscription && subscription.remove();
-    setSubscription(null);
-  };
-
-  useEffect(() => {
-    _subscribe();
-    return () => _unsubscribe();
-  }, []);
+  const isShaking = useIsShaking();
 
   useEffect(() => {
     async function getRandomQuote() {
