+++ utils/utils.js
@@ -1,0 +1,32 @@
+import React, { useState, useEffect } from "react";
+import { Accelerometer } from "expo-sensors";
+
+function useIsShaking() {
+  const [isShaking, setIsShaking] = useState(false);
+  const [subscription, setSubscription] = useState(null);
+
+  Accelerometer.setUpdateInterval(400);
+
+  const _subscribe = () => {
+    setSubscription(
+      Accelerometer.addListener((accelerometerData) => {
+        const { x, y, z } = accelerometerData;
+        setIsShaking(Math.abs(x) + Math.abs(y) + Math.abs(z) > 1.78);
+      })
+    );
+  };
+
+  const _unsubscribe = () => {
+    subscription && subscription.remove();
+    setSubscription(null);
+  };
+
+  useEffect(() => {
+    _subscribe();
+    return () => _unsubscribe();
+  }, []);
+
+  return isShaking;
+}
+
+export { useIsShaking };
